+++ example/android/app/src/main/kotlin/es/antonborri/home_widget_example/HomeWidgetExampleProvider.kt
@@ -3,7 +3,9 @@ package es.antonborri.home_widget_example
 import android.appwidget.AppWidgetManager
 import android.content.Context
 import android.content.SharedPreferences
+import android.graphics.BitmapFactory
 import android.net.Uri
+import android.view.View
 import android.widget.RemoteViews
 import es.antonborri.home_widget.HomeWidgetBackgroundIntent
 import es.antonborri.home_widget.HomeWidgetLaunchIntent
@@ -32,6 +34,15 @@ class HomeWidgetExampleProvider : HomeWidgetProvider() {
                 val message = widgetData.getString("message", null)
                 setTextViewText(R.id.widget_message, message
                         ?: "No Message Set")
+                // Show Images saved with `renderFlutterWidget`
+                val image = widgetData.getString("dashIcon", null)
+                if (image != null) {
+                 setImageViewBitmap(R.id.widget_img, BitmapFactory.decodeFile(image))
+                 setViewVisibility(R.id.widget_img, View.VISIBLE)
+                } else {
+                    setViewVisibility(R.id.widget_img, View.GONE)
+                }
+
                 // Detect App opened via Click inside Flutter
                 val pendingIntentWithData = HomeWidgetLaunchIntent.getActivity(
                         context,